refactor(core): the config lock moves where both the daemon and the CLI can hold it
