+++ crates/shep-cli/src/commands/shep_toml.rs
@@ -27,11 +27,10 @@
 //! Phase 15 made `shep` a library with three thin `[[bin]]` targets over
 //! it. `flock(2)` and unix mode bits are Unix-only, so each call site that
 //! needs one gates itself individually and carries a real Windows arm
-//! rather than an unimplemented one: [`ConfigLock`] and
-//! [`ConfigLock::acquire`] below are each `#[cfg(unix)]`/`#[cfg(windows)]`
-//! pairs (`flock(2)` against `share_mode(0)`), and [`create_home_dir`]
-//! folds a unix-only `.mode()` call into an otherwise portable
-//! `DirBuilder`.
+//! rather than an unimplemented one: `shep_core::config_lock::ConfigLock`
+//! and its `acquire` are each `#[cfg(unix)]`/`#[cfg(windows)]` pairs
+//! (`flock(2)` against `share_mode(0)`), and [`create_home_dir`] folds a
+//! unix-only `.mode()` call into an otherwise portable `DirBuilder`.
 
 // `clippy::result_large_err` fires on every `Result<_, ShepTomlError>`
 // signature in this module on Windows, and on none of them on macOS or
@@ -51,7 +50,7 @@
 use std::collections::BTreeMap;
 use std::io::Write as _;
 #[cfg(unix)]
-use std::os::unix::fs::{DirBuilderExt as _, OpenOptionsExt as _};
+use std::os::unix::fs::DirBuilderExt as _;
 use std::path::{Path, PathBuf};
 
 use toml_edit::{Array, DocumentMut, Item, Table, Value};
@@ -865,120 +864,11 @@ fn create_home_dir(dir: &Path) -> std::io::Result<()> {
     builder.create(dir)
 }
 
-/// Creates the staging file a config is written through, in `parent` so
-/// the later `rename` stays within one filesystem.
-///
-/// The create-at-mode reasoning lives with
-/// [`shep_core::atomic_file::create_staging_file`], which four stores now
-/// share. What is left here is the pair of names, and this wrapper is
-/// where they stay: `commands::dog_migration` writes `dogs.toml` through
-/// the same staging name as `shep.toml`, and two call sites spelling that
-/// pair out separately is how the two would drift.
-pub(super) fn create_config_file(parent: &Path) -> std::io::Result<tempfile::NamedTempFile> {
-    shep_core::atomic_file::create_staging_file(parent, "shep", ".toml.tmp")
-}
-
-/// An exclusive advisory lock over one config file, held for as long as
-/// the value lives and released when it drops (including on an early `?`,
-/// and by the kernel if the process dies holding it).
-///
-/// Keyed on the path it is given rather than on `shep.toml` specifically:
-/// [`ShepToml::edit`] takes one over `shep.toml`, and
-/// `commands::dog_migration` takes one over `dogs.toml`, which has two
-/// writers of its own. **Whenever both are held at once, `shep.toml`'s is
-/// taken first**, which is the whole of what keeps the two orderings from
-/// deadlocking; `migrate_dog_sections` is the one caller that holds both,
-/// and it says so at the point it nests them.
-///
-/// The lock is on a sibling `<name>.lock`, never on the config itself,
-/// and that is the whole design decision — the same one `barks::RingLock`
-/// records: [`ShepToml::save`] finishes by `rename`ing a new file over the
-/// config, which replaces the inode. A lock taken on the config would be a
-/// lock on an inode the very next successful save unlinks; the next writer
-/// would open the *new* inode, find it unlocked, and the two would be
-/// excluding nothing. The lock file is never renamed, never rewritten and
-/// never read; it exists only to be an inode with a stable identity, and
-/// it is left on disk between edits on purpose so both writers keep
-/// agreeing on which one it is.
-pub(super) struct ConfigLock {
-    /// `flock(2)` is released by this handle's `Drop`. Named with a
-    /// leading underscore because it is held, never read.
-    #[cfg(unix)]
-    _flock: nix::fcntl::Flock<std::fs::File>,
-    /// The lock file, opened with `share_mode(0)`. The same primitive and
-    /// the same sibling-file shape [`shep_core::kv`] and
-    /// [`shep_core::barks`] use; see either for the full argument.
-    #[cfg(windows)]
-    _handle: std::fs::File,
-}
-
-impl ConfigLock {
-    /// Blocks until this process holds `path`'s lock exclusively.
-    ///
-    /// # Errors
-    /// The lock file could not be created beside `path`, or `flock` failed
-    /// for a reason other than contention (contention blocks rather than
-    /// failing).
-    #[cfg(windows)]
-    pub(super) fn acquire(path: &Path) -> std::io::Result<Self> {
-        use std::os::windows::fs::OpenOptionsExt as _;
-
-        /// Another handle already holds share access this open denies.
-        const ERROR_SHARING_VIOLATION: i32 = 32;
-        /// How long a contended retry sleeps. The unix arm blocks in the
-        /// kernel; this polls, for the reason `shep_core::kv` documents.
-        const RETRY_INTERVAL: std::time::Duration = std::time::Duration::from_millis(2);
-
-        let lock_path = lock_path(path);
-        loop {
-            match std::fs::OpenOptions::new()
-                .write(true)
-                .create(true)
-                .truncate(false)
-                .share_mode(0)
-                .open(&lock_path)
-            {
-                Ok(handle) => return Ok(Self { _handle: handle }),
-                Err(error) if error.raw_os_error() == Some(ERROR_SHARING_VIOLATION) => {
-                    std::thread::sleep(RETRY_INTERVAL);
-                }
-                Err(error) => return Err(error),
-            }
-        }
-    }
-
-    #[cfg(unix)]
-    pub(super) fn acquire(path: &Path) -> std::io::Result<Self> {
-        use nix::fcntl::{Flock, FlockArg};
-
-        let file = std::fs::OpenOptions::new()
-            .write(true)
-            .create(true)
-            .truncate(false)
-            .mode(shep_core::atomic_file::OWNER_ONLY_FILE_MODE)
-            .open(lock_path(path))?;
-
-        // `LockExclusive` blocks; the non-blocking variant would need a
-        // retry loop and a deadline, and a `shep enable` that waits its
-        // turn behind a concurrent `shep adopt` is exactly the behaviour
-        // wanted here.
-        Flock::lock(file, FlockArg::LockExclusive)
-            .map(|flock| Self { _flock: flock })
-            .map_err(|(_file, errno)| std::io::Error::from(errno))
-    }
-}
-
-/// The lock file that guards `path`: its own name with `.lock` appended,
-/// so it sits in `$SHEP_HOME` next to the config and inherits that
-/// directory's `0700`.
-fn lock_path(path: &Path) -> PathBuf {
-    let mut name = path
-        .file_name()
-        .map(std::ffi::OsStr::to_os_string)
-        .unwrap_or_default();
-    name.push(".lock");
-    path.parent().unwrap_or_else(|| Path::new(".")).join(name)
-}
+/// `ConfigLock` and `create_config_file` moved to
+/// [`shep_core::config_lock`] so shep-daemon can hold the same lock over
+/// `dogs.toml` shep-cli does; both names still resolve here unchanged for
+/// this crate's own callers.
+pub(super) use shep_core::config_lock::{ConfigLock, create_config_file};
 
 /// What [`ShepToml::edit`] can fail with. Module-scoped per IR-18.
 ///
